feature : 틸틸이 티어 기능 구현 완료

+++ server/src/main/java/com/codestates/TILTILE/member/dto/MemberWithBookmarksDto.java
@@ -13,7 +13,7 @@ public class MemberWithBookmarksDto {
     private Long memberId;
     private String nickName;
 
-    private long tilTier;
+    private Integer tilTier;
 
     private String aboutMe;
